Added Part V – User Authentication

+++ controllers/product.js
@@ -16,6 +16,7 @@ module.exports.addPost = (req, resp) => {
 
     let productFromForm = req.body;
     productFromForm.image = '/' + req.file.path;
+    productFromForm.creator = req.user._id;
 
     Product.create(productFromForm).then((insertedProduct) => {
 
@@ -32,7 +33,7 @@ module.exports.addPost = (req, resp) => {
     }).catch(console.log);
 };
 
-module.exports.editGet = (req, resp) => {
+module.exports.editGet = (req, res) => {
 
     let id = req.params.id;
 
@@ -43,14 +44,30 @@ module.exports.editGet = (req, resp) => {
             return;
         }
 
-        Category.find().then((categories) => {
+        if (product.creator.equals(req.user._id) || req.user.roles.indexOf('Admin') > -1) {
 
-            resp.render('products/edit', {product: product, categories: categories});
-        });
+            Category.find().then((categories) => {
+                res.render('product/edit', {
+                    product: product,
+                    categories: categories
+                });
+            }).catch((err) => {
+                console.log(err);
+                return;
+
+            });
+        } else {
+            res.render('error');
+            return;
+        }
+    }).catch((err) => {
+        res.render('error');
+        console.log(err);
+        return;
     });
 };
 
-module.exports.editPost = (req, resp) => {
+module.exports.editPost = (req, res) => {
 
     let id = req.params.id;
     let editedProduct = req.body;
@@ -58,7 +75,7 @@ module.exports.editPost = (req, resp) => {
     Product.findById(id).then((product) => {
 
         if (!product) {
-            resp.redirect(`/?error=${encodeURIComponent('error=Product was not found!')}`);
+            res.redirect(`/?error=${encodeURIComponent('error=Product was not found!')}`);
             return;
         }
 
@@ -78,9 +95,8 @@ module.exports.editPost = (req, resp) => {
 
                     let index = currentCategory.products.indexOf(product._id);
 
-                    if (index >= 0) {
-
-                        currentCategory.products.splice(index, 1);
+                    if (index > 0) {
+                        currentCategory.product.splice(index, 1);
                     }
 
                     currentCategory.save();
@@ -89,20 +105,31 @@ module.exports.editPost = (req, resp) => {
                     nextCategory.save();
 
                     product.category = editedProduct.category;
-
                     product.save().then(() => {
 
-                        resp.redirect('/?success=' + encodeURIComponent('Product was edited successfully!'));
+                        res.redirect(`/?success= ${encodeURIComponent('Product was edited successfully!')}`);
+
+                    }).catch((err) => {
+                        console.log(err);
+                        return;
                     });
+                }).catch((err) => {
+                    console.log(err);
+                    return;
                 });
             });
         } else {
-
             product.save().then(() => {
+                res.redirect(`/?success= ${encodeURIComponent('Product was edited successfully!')}`);
 
-                resp.redirect('/?success=' + encodeURIComponent('Product was edited successfully!'));
+            }).catch((err) => {
+                console.log(err);
+                return;
             });
         }
+    }).catch((err) => {
+        console.log(err);
+        return;
     });
 };
 
@@ -113,14 +140,21 @@ module.exports.deleteGet = (req, res) => {
     Product.findById(id).then((product) => {
 
         if (!product) {
+
             res.sendStatus(404);
             return;
         }
 
-        res.render('products/delete', {product: product});
+        if (product.creator.equals(req.user._id) || req.user.roles.indexOf('Admin') > -1) {
 
+            res.render('product/delete', { product: product });
+        } else {
+            res.render('error');
+            return;
+        }
     }).catch((err) => {
 
+        res.render('error');
         console.log(err);
         return;
     });
@@ -138,11 +172,13 @@ module.exports.deletePost = (req, res) => {
             return;
         }
 
-        fs.unlink(path.normalize(path.join('.', product.image)), () => {
+        if (product.creator.equals(req.user._id) || req.user.roles.indexOf('Admin') > -1) {
 
-            res.redirect(`/?success= ${encodeURIComponent('Product was deleted successfully!')}`);
-        });
+            fs.unlink(path.normalize(path.join('.', product.image)), () => {
 
+                res.redirect(`/?success= ${encodeURIComponent('Product was deleted successfully!')}`);
+            });
+        }
     }).catch((err) => {
         console.log(err);
         return;
@@ -162,4 +198,29 @@ module.exports.buyGet = (req, res) => {
 
         res.render('products/buy', { product: product });
     });
+};
+
+module.exports.buyPost = (req, res) => {
+
+  let productId = req.params.id;
+
+  Product.findById(productId).then(product => {
+
+     if (product.buyer) {
+
+         let error = `error=${encodeURIComponent('Product was already bought!')}`;
+         res.redirect(`/${error}`);
+         return;
+     }
+
+     product.buyer = req.user._id;
+     product.save().then(() => {
+
+         req.user.boughtProducts.push(productId);
+         req.user.save().then(() => {
+
+             res.redirect('/');
+         });
+     });
+  });
 };
